+++ src/image.rs
@@ -5,12 +5,16 @@ use std::mem::size_of;
 use std::ops::{Deref, DerefMut};
 use std::path::Path;
 
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
 use crate::color::{ColorType, ExtendedColorType};
 use crate::error::{
     ImageError, ImageFormatHint, ImageResult, LimitError, LimitErrorKind, ParameterError,
-    ParameterErrorKind,
+    ParameterErrorKind, UnsupportedError, UnsupportedErrorKind,
 };
 use crate::math::Rect;
+use crate::metadata::Orientation;
 use crate::traits::Pixel;
 use crate::ImageBuffer;
 
@@ -19,6 +23,7 @@ use crate::animation::Frames;
 /// An enumeration of supported image formats.
 /// Not all formats support both encoding and decoding.
 #[derive(Clone, Copy, PartialEq, Eq, Debug, Hash)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[non_exhaustive]
 pub enum ImageFormat {
     /// An Image in PNG Format
@@ -65,6 +70,9 @@ pub enum ImageFormat {
 
     /// An Image in QOI Format
     Qoi,
+
+    /// An Image in PCX Format
+    Pcx,
 }
 
 impl ImageFormat {
@@ -89,7 +97,7 @@ impl ImageFormat {
 
             Some(match ext.as_str() {
                 "avif" => ImageFormat::Avif,
-                "jpg" | "jpeg" => ImageFormat::Jpeg,
+                "jpg" | "jpeg" | "jfif" => ImageFormat::Jpeg,
                 "png" | "apng" => ImageFormat::Png,
                 "gif" => ImageFormat::Gif,
                 "webp" => ImageFormat::WebP,
@@ -103,6 +111,7 @@ impl ImageFormat {
                 "pbm" | "pam" | "ppm" | "pgm" => ImageFormat::Pnm,
                 "ff" => ImageFormat::Farbfeld,
                 "qoi" => ImageFormat::Qoi,
+                "pcx" => ImageFormat::Pcx,
                 _ => return None,
             })
         }
@@ -178,6 +187,7 @@ impl ImageFormat {
             // Qoi's MIME type is being worked on.
             // See: https://github.com/phoboslab/qoi/issues/167
             "image/x-qoi" => Some(ImageFormat::Qoi),
+            "image/vnd.zbrush.pcx" | "image/x-pcx" => Some(ImageFormat::Pcx),
             _ => None,
         }
     }
@@ -225,6 +235,7 @@ impl ImageFormat {
             ImageFormat::Qoi => "image/x-qoi",
             // farbfeld's MIME type taken from https://www.wikidata.org/wiki/Q28206109
             ImageFormat::Farbfeld => "application/octet-stream",
+            ImageFormat::Pcx => "image/vnd.zbrush.pcx",
         }
     }
 
@@ -249,6 +260,7 @@ impl ImageFormat {
             ImageFormat::Farbfeld => true,
             ImageFormat::Avif => true,
             ImageFormat::Qoi => true,
+            ImageFormat::Pcx => true,
         }
     }
 
@@ -273,6 +285,7 @@ impl ImageFormat {
             ImageFormat::OpenExr => true,
             ImageFormat::Dds => false,
             ImageFormat::Qoi => true,
+            ImageFormat::Pcx => false,
         }
     }
 
@@ -304,6 +317,7 @@ impl ImageFormat {
             // According to: https://aomediacodec.github.io/av1-avif/#mime-registration
             ImageFormat::Avif => &["avif"],
             ImageFormat::Qoi => &["qoi"],
+            ImageFormat::Pcx => &["pcx"],
         }
     }
 
@@ -326,6 +340,7 @@ impl ImageFormat {
             ImageFormat::Farbfeld => cfg!(feature = "ff"),
             ImageFormat::Avif => cfg!(feature = "avif"),
             ImageFormat::Qoi => cfg!(feature = "qoi"),
+            ImageFormat::Pcx => cfg!(feature = "pcx"),
             ImageFormat::Dds => false,
         }
     }
@@ -349,6 +364,7 @@ impl ImageFormat {
             ImageFormat::OpenExr => cfg!(feature = "exr"),
             ImageFormat::Qoi => cfg!(feature = "qoi"),
             ImageFormat::Hdr => cfg!(feature = "hdr"),
+            ImageFormat::Pcx => false,
             ImageFormat::Dds => false,
         }
     }
@@ -371,6 +387,7 @@ impl ImageFormat {
             ImageFormat::Qoi,
             ImageFormat::Dds,
             ImageFormat::Hdr,
+            ImageFormat::Pcx,
         ]
         .iter()
         .copied()
@@ -636,6 +653,17 @@ pub trait ImageDecoder {
         Ok(None)
     }
 
+    /// Returns the orientation of the image.
+    ///
+    /// This is usually obtained from the Exif metadata, if present. Formats that don't support
+    /// indicating orientation in their image metadata will return `Ok(Orientation::NoTransforms)`.
+    fn orientation(&mut self) -> ImageResult<Orientation> {
+        Ok(self
+            .exif_metadata()?
+            .and_then(|chunk| Orientation::from_exif_chunk(&chunk))
+            .unwrap_or(Orientation::NoTransforms))
+    }
+
     /// Returns the total number of bytes in the decoded image.
     ///
     /// This is the size of the buffer that must be passed to `read_image` or
@@ -787,6 +815,25 @@ pub trait ImageEncoder {
         height: u32,
         color_type: ExtendedColorType,
     ) -> ImageResult<()>;
+
+    /// Set the ICC profile to use for the image.
+    ///
+    /// This function is a no-op for formats that don't support ICC profiles.
+    /// For formats that do support ICC profiles, the profile will be embedded
+    /// in the image when it is saved.
+    ///
+    /// # Errors
+    ///
+    /// This function returns an error if the format does not support ICC profiles.
+    fn set_icc_profile(&mut self, icc_profile: Vec<u8>) -> Result<(), UnsupportedError> {
+        let _ = icc_profile;
+        Err(UnsupportedError::from_format_and_kind(
+            ImageFormatHint::Unknown,
+            UnsupportedErrorKind::GenericFeature(
+                "ICC profiles are not supported for this format".into(),
+            ),
+        ))
+    }
 }
 
 /// Immutable pixel iterator
@@ -799,7 +846,7 @@ pub struct Pixels<'a, I: ?Sized + 'a> {
     height: u32,
 }
 
-impl<'a, I: GenericImageView> Iterator for Pixels<'a, I> {
+impl<I: GenericImageView> Iterator for Pixels<'_, I> {
     type Item = (u32, u32, I::Pixel);
 
     fn next(&mut self) -> Option<(u32, u32, I::Pixel)> {
@@ -833,7 +880,7 @@ impl<I: ?Sized> Clone for Pixels<'_, I> {
 /// use image::{GenericImageView, Rgb, RgbImage};
 ///
 /// let buffer = RgbImage::new(10, 10);
-/// let image: &dyn GenericImageView<Pixel=Rgb<u8>> = &buffer;
+/// let image: &dyn GenericImageView<Pixel = Rgb<u8>> = &buffer;
 /// ```
 pub trait GenericImageView {
     /// The type of pixel.
@@ -1222,6 +1269,7 @@ where
     I: Deref,
 {
     type Target = SubImageInner<I>;
+
     fn deref(&self) -> &Self::Target {
         &self.inner
     }
@@ -1616,6 +1664,7 @@ mod tests {
         assert_eq!(from_path("./a.Ppm").unwrap(), ImageFormat::Pnm);
         assert_eq!(from_path("./a.pgm").unwrap(), ImageFormat::Pnm);
         assert_eq!(from_path("./a.AViF").unwrap(), ImageFormat::Avif);
+        assert_eq!(from_path("./a.PCX").unwrap(), ImageFormat::Pcx);
         assert!(from_path("./a.txt").is_err());
         assert!(from_path("./a").is_err());
     }
@@ -1771,7 +1820,7 @@ mod tests {
     fn image_formats_are_recognized() {
         use ImageFormat::*;
         const ALL_FORMATS: &[ImageFormat] = &[
-            Avif, Png, Jpeg, Gif, WebP, Pnm, Tiff, Tga, Dds, Bmp, Ico, Hdr, Farbfeld, OpenExr,
+            Avif, Png, Jpeg, Gif, WebP, Pnm, Tiff, Tga, Dds, Bmp, Ico, Hdr, Farbfeld, OpenExr, Pcx,
         ];
         for &format in ALL_FORMATS {
             let mut file = Path::new("file.nothing").to_owned();