+++ src/lib.rs
@@ -56,13 +56,13 @@
 //! The two main types for storing images:
 //! * [`ImageBuffer`] which holds statically typed image contents.
 //! * [`DynamicImage`] which is an enum over the supported `ImageBuffer` formats
-//!     and supports conversions between them.
+//!   and supports conversions between them.
 //!
 //! As well as a few more specialized options:
 //! * [`GenericImage`] trait for a mutable image buffer.
 //! * [`GenericImageView`] trait for read only references to a `GenericImage`.
 //! * [`flat`] module containing types for interoperability with generic channel
-//!     matrices and foreign interfaces.
+//!   matrices and foreign interfaces.
 //!
 //! [`GenericImageView`]: trait.GenericImageView.html
 //! [`GenericImage`]: trait.GenericImage.html
@@ -119,6 +119,14 @@
 #![deny(missing_copy_implementations)]
 #![cfg_attr(all(test, feature = "benchmarks"), feature(test))]
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
+// We've temporarily disabled PCX support for 0.25.5 release
+// by removing the corresponding feature.
+// We want to ship bug fixes without committing to PCX support.
+//
+// Cargo shows warnings about code depending on a nonexistent feature
+// even to people using the crate as a dependency,
+// so we have to suppress those warnings.
+#![allow(unexpected_cfgs)]
 
 #[cfg(all(test, feature = "benchmarks"))]
 extern crate test;
@@ -173,7 +181,6 @@ pub use crate::image_reader::{ImageReader, LimitSupport, Limits};
 pub use crate::dynimage::DynamicImage;
 
 pub use crate::animation::{Delay, Frame, Frames};
-pub use crate::metadata::Orientation;
 
 // More detailed error type
 pub mod error;
@@ -207,7 +214,7 @@ pub mod flat;
 ///
 /// | Format   | Decoding                                  | Encoding                                |
 /// | -------- | ----------------------------------------- | --------------------------------------- |
-/// | AVIF     | Yes (8-bit only) \*                       | Yes (lossy only)                        |
+/// | AVIF     | Yes \*                                    | Yes (lossy only)                        |
 /// | BMP      | Yes                                       | Yes                                     |
 /// | DDS      | Yes                                       | ---                                     |
 /// | Farbfeld | Yes                                       | Yes                                     |
@@ -265,6 +272,8 @@ pub mod codecs {
     pub mod jpeg;
     #[cfg(feature = "exr")]
     pub mod openexr;
+    #[cfg(feature = "pcx")]
+    pub mod pcx;
     #[cfg(feature = "png")]
     pub mod png;
     #[cfg(feature = "pnm")]
@@ -291,7 +300,7 @@ mod color;
 mod dynimage;
 mod image;
 mod image_reader;
-mod metadata;
+pub mod metadata;
 //TODO delete this module after a few releases
 /// deprecated io module the original io module has been renamed to `image_reader`
 pub mod io {
@@ -320,7 +329,7 @@ mod utils;
 macro_rules! insert_as_doc {
     { $content:expr } => {
         #[allow(unused_doc_comments)]
-        #[doc = $content] extern { }
+        #[doc = $content] extern "Rust" { }
     }
 }