+++ CHANGES.md
@@ -1,11 +1,42 @@
 # Release Notes
 
 ## Known issues
-- Many decoders will panic on malicous input. In most cases, this is caused by
-  not enforcing memory limits, though other panics have been seen from fuzzing.
+- Many decoders will panic on malicious input.
 - The color space information of pixels is not clearly communicated.
 
 ## Changes
+
+### Version 0.25.5
+
+Features:
+ - Added support for decoding 10-bit and 12-bit AVIF
+ - Initial, opt-in serde support for an enum. This may be extended to other types in the future.
+
+Bug fixes:
+ - [Multiple bug fixes in AVIF decoding](https://github.com/image-rs/image/pull/2373)
+ - The `rayon` feature now correctly toggles the use of `rayon` when encoding AVIF. (Previously it would be either always on or always off depending on the version of the `ravif` crate in your dependency tree.)
+ - "jfif" file extension for JPEG images is now recognized
+
+### Version 0.25.4
+
+Features:
+ - Much faster decoding of lossless WebP due to a variety of optimizations. Our benchmarks show 2x to 2.5x improvement.
+ - Added support for orientation metadata, so that e.g. smartphone camera images could be displayed correctly:
+   - Added `ImageDecoder::orientation()` and implemented orientation metadata extraction for JPEG, WebP and TIFF formats
+   - Added `DynamicImage::apply_orientation()` to apply the orientation to an image
+ - Added support for extracting Exif metadata from images via `ImageDecoder::exif_metadata()`, and implemented it for JPEG and WebP formats
+ - Added `ImageEncoder::set_icc_profile()` and implemented it for WebP format. Pull requests with implementations for other formats are welcome.
+ - Added `DynamicImage::fast_blur()` for a linear-time approximation of Gaussian blur, which is much faster at larger blur radii
+
+Bug fixes:
+ - Fixed some APNG images being decoded incorrectly
+ - Fixed the iterator over animated WebP frames to return `None` instead of an error when the end of the animation is reached
+
+### Version 0.25.3
+
+Yanked! This version accidentally missed a commit that should have been
+included with the release. The `Orientation` struct should be in the
+appropriate module instead of the top-level. This release won't be supported.
 
 ### Version 0.25.2